+++ CHANGELOG.md
@@ -14,6 +14,7 @@ For a full diff see [`7039e81...main`][7039e81...main].
 - Implemented `Expression\Value` as a value object ([#2]), by [@localheinz]
 - Implemented `Expression\Assignment` as a value object ([#3]), by [@localheinz]
 - Implemented `Converter\NullFrontMatterConverter` ([#4]), by [@localheinz]
+- Implemented `Converter\ToSingleAssignmentFrontMatterConverter` ([#5]), by [@localheinz]
 
 [7039e81...main]: https://github.com/ergebnis/twig-front-matter/compare/7039e81...main
 
@@ -21,5 +22,6 @@ For a full diff see [`7039e81...main`][7039e81...main].
 [#2]: https://github.com/ergebnis/twig-front-matter/pull/2
 [#3]: https://github.com/ergebnis/twig-front-matter/pull/3
 [#4]: https://github.com/ergebnis/twig-front-matter/pull/4
+[#5]: https://github.com/ergebnis/twig-front-matter/pull/5
 
 [@localheinz]: https://github.com/localheinz